core: move benchmarks into bench_test.go
They share some functions and they're benchmarks, so let's have them here.
